more refactoring of akk code
